Fixes warnings on Windows (#288)

+++ lib/src/legacy/legacy_h26x_auth.c
@@ -117,7 +117,7 @@ legacy_verify_hashes_with_hash_list(legacy_sv_t *self,
   assert(hash_size > 0);
   // Expected hashes.
   uint8_t *expected_hashes = self->gop_info->hash_list;
-  const int num_expected_hashes = self->gop_info->list_idx / hash_size;
+  const int num_expected_hashes = (const int)(self->gop_info->list_idx / hash_size);
 
   legacy_h26x_nalu_list_t *nalu_list = self->nalu_list;
   legacy_h26x_nalu_list_item_t *last_used_item = NULL;
